refactor: cut out the getConfigFilePath function to a separate file
